v0.91g r45

+++ dscoverloader/source/main.c
@@ -55,7 +55,7 @@ void Main(){
 		//ucs2tombs(target,extlink.DataFullPathFilenameUnicode);
 			strcpy(target,filename);goto target_set;
 	}else if(!strcmp((char*)head+0x1e0,"mshl2wrap link")){
-		unsigned int s=(head[0x1f0]<<24)+(head[0x1f1]<<16)+(head[0x1f2]<<8)+head[0x1f3];
+		unsigned int s=((unsigned int)head[0x1f0]<<24)+(head[0x1f1]<<16)+(head[0x1f2]<<8)+head[0x1f3];
 		_consolePrint("Detected mshl2wrap link.\n");
 		if(size<s+256*3){fclose(f);_consolePrint("Failed.\n");die();}
 		fseek(f,s,SEEK_SET);fread(target,1,256*3,f);goto target_set;
